requirements.txt
asgiref==3.6.0
cement==2.8.2
certifi==2023.5.7
charset-normalizer==2.0.12
colorama==0.4.3
Django==4.2
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.7
django-ipware==5.0.0
django-session-timeout==0.1.0
idna==3.4
jmespath==1.0.1
pathspec==0.10.1
Pillow==9.5.0
pypiwin32==223
python-decouple==3.8
pywin32==306
PyYAML==5.4.1
requests==2.30.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
termcolor==1.1.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.1.9
